Got everything working. This is good to go for production.
