+++ TCGTradingPost/src/main/resources/static/css/cart.css
@@ -1,4 +1,4 @@
-#carts_container {
+#cart_container {
 	z-index: 0;
 	position: absolute;
 	margin-top: 200px;
