+++ TCGTradingPost/src/main/resources/static/js/cardSearch.js
@@ -5,8 +5,22 @@ let xhrResultsArr = [];
 let searchApp = angular.module("searchApp", []);
 searchApp.controller("searchCtrl", function($scope) {
 	$scope.apiSearchResults = searchResultsArr;
-	$scope.buyBtnFunc = function() {
+	$scope.buyBtnFunc = function(cardID) {
 		
+		let formData = new FormData();
+		formData.append("cardID", cardID);
+		
+		let xhr = new XMLHttpRequest();
+		xhr.onreadystatechange = function() {
+			console.log("readystate: "+xhr.readyState+" status: "+xhr.status)
+			if(xhr.readyState === 4 && xhr.status === 200) {
+				
+				alert(xhr.responseText);
+				
+			}
+		}
+		xhr.open("POST", "/buyItem", true);
+		xhr.send(formData);
 	}
 	$scope.searchResultsFunc = function(xhttpr) {
 		console.log("searchResultsFunc");
@@ -170,20 +184,3 @@ searchApp.controller("searchCtrl", function($scope) {
 });
 
 
-let buyBtnFunc = function(cardID) {
-	
-	let formData = new FormData();
-	formData.append("cardID", cardID);
-	
-	let xhr = new XMLHttpRequest();
-	xhr.onreadystatechange = function() {
-		console.log("readystate: "+xhr.readyState+" status: "+xhr.status)
-		if(xhr.readyState === 4 && xhr.status === 200) {
-			
-			alert(xhr.responseText);
-			
-		}
-	}
-	xhr.open("POST", "/buyItem", true);
-	xhr.send(formData);
-}
